Change Money input type to Rational
The whole aim was to make constructing Money values with literals work.
The compiler has been persistent with warning about defaults; in this
case, Ratio Integer aka Rational is a better choice than messing with
the inprecision of Doubles (which already burned us once). And it's
money, so why not. Billions and billions!